feat: Implement Seccomp User Notification bridge for HAL proxying
- Introduces Seccomp Bridge architecture for kernel-level syscall emulation.
- Uses SECCOMP_RET_USER_NOTIF to intercept specific ioctl commands.
- Implements asynchronous listener thread in the monitor process.
- Performs zero-copy memory writes to container via process_vm_writev.
- Includes robust compatibility blocks for pre-5.0 kernels and older headers.
- Highly optimized BPF filter targeting only specific bridge ioctls.
- Fixes ioctl request overflow for 64-bit musl systems.
- Adds virtual hardware stub at /dev/ds-bridge.

+++ src/container.c
@@ -172,7 +172,7 @@ static void *seccomp_bridge_listener(void *arg) {
 
   while (1) {
     memset(&req, 0, sizeof(req));
-    if (ioctl(fd, (unsigned long)SECCOMP_IOCTL_NOTIF_RECV, &req) < 0) {
+    if (ioctl(fd, (int)(uintptr_t)SECCOMP_IOCTL_NOTIF_RECV, &req) < 0) {
       if (errno == EINTR)
         continue;
       break;
@@ -205,7 +205,7 @@ static void *seccomp_bridge_listener(void *arg) {
       resp.flags = SECCOMP_USER_NOTIF_FLAG_CONTINUE;
     }
 
-    if (ioctl(fd, (unsigned long)SECCOMP_IOCTL_NOTIF_SEND, &resp) < 0) {
+    if (ioctl(fd, (int)(uintptr_t)SECCOMP_IOCTL_NOTIF_SEND, &resp) < 0) {
       if (errno == EINTR || errno == ENOENT)
         continue;
       break;
